pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "neuromancer"
authors = [
    { name = "Aaron Tuor", email = "aaron.tuor@pnnl.gov" },
    { name = "Jan Drgona", email = "jan.drgona@pnnl.gov" },
    { name = "James Koch", email = "james.koch@pnnl.gov" },
    { name = "Madelyn Shapiro", email = "madelyn.shapiro@pnnl.gov"},
    { name = "Draguna Vrabie", email = "draguna.vrabie@pnnl.gov" }
]
description = "Neural Modules with Adaptive Nonlinear Constraints and Efficient Regularization"
readme = "README.md"
requires-python = ">=3.9, <3.11"
keywords = [
    "Deep Learning",
    "Pytorch",
    "Linear Models",
    "Dynamical Systems",
    "Data-driven control",
]



license = { text = "BSD-2-Clause" }
classifiers = [
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Physics",
    "License :: OSI Approved :: BSD License"
]
dependencies = [
    "dill",
    "graphviz",
    "matplotlib",
    "mlflow==2.5.0",
    "networkx==3.0.0",
    "numpy<1.24.0",
    "pandas",
    "plum_dispatch==1.7.3",
    "scikit_learn",
    "scipy",
    "six",
    "pydot==1.4.2",
    "pyts",
    "torch",
    "torchdiffeq",
    "toml"
]

version = "1.4.2"

[project.urls]
homepage = "https://github.com/pnnl/neuromancer/"
documentation = "https://github.com/pnnl/neuromancer/"
repository = "https://github.com/pnnl/neuromancer/"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
tests = ["pytest", "hypothesis"]
examples = ["casadi", "cvxpy", "imageio", "cvxpylayers"]
docs = ["sphinx", "sphinx-rtd-theme"]